feat(ArticleDetail): extract SetPaymentPasswordForm
