+++ src/components/Forms/index.tsx
@@ -4,5 +4,6 @@ export * from './CreateCircleForm'
 export * from './EmailLoginForm'
 export * from './EmailSignUpForm'
 export * from './SelectAuthMethodForm'
+export * from './SetPaymentPasswordForm'
 export * from './Verification'
 export * from './WalletAuthForm'
